lxd: split cluster group endpoints into api_cluster_group.go
- Move /1.0/cluster/groups* endpoint definitions, handlers, and
group-specific helpers into api_cluster_group.go.
